services/douyin-adapter/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0
loguru==0.7.2
cryptography==41.0.7
